ai-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pillow==10.2.0
numpy==1.26.3
opencv-python==4.9.0.80
scikit-learn==1.4.0
torch==2.1.2
torchvision==0.16.2
python-dotenv==1.0.0
pydantic==2.5.3
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pgvector==0.2.4

# For face recognition (optional)
# face-recognition==1.3.0
# dlib==19.24.2
